move `IProjectUpdatedHandler` implementation to its own type
