+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/ProjectSystem/LanguageServices/Handlers/SourceItemHandler_CommandLineTests.cs
@@ -23,7 +23,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.LanguageServices.Handlers
 
             Assert.Throws<ArgumentNullException>("project", () =>
             {
-                new SourceItemHandler((UnconfiguredProject)null, null);
+                new SourceItemHandler((UnconfiguredProject)null);
             });
         }
 
@@ -87,7 +87,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.LanguageServices.Handlers
         {
             project = project ?? UnconfiguredProjectFactory.Create();
 
-            var handler = new SourceItemHandler(project, null);
+            var handler = new SourceItemHandler(project);
             if (context != null)
                 handler.Initialize(context);
